+++ lemon-development/doc/html/search/groups_6.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['heap_20structures_74',['Heap Structures',['../group__heaps.html',1,'']]]
-];