+++ lemon-development/doc/html/search/groups_7.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['input_2doutput_75',['Input-Output',['../group__io__group.html',1,'']]]
-];