+++ lemon-development/doc/html/search/groups_a.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['nauty_20format_90',['NAUTY Format',['../group__nauty__group.html',1,'']]]
-];