+++ lemon-development/doc/html/search/pages_0.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['lemon_20coding_20style_100',['LEMON Coding Style',['../coding_style.html',1,'']]],
-  ['lemon_20graph_20format_20_28lgf_29_101',['LEMON Graph Format (LGF)',['../lgf-format.html',1,'']]],
-  ['license_20terms_102',['License Terms',['../license.html',1,'']]]
-];